move mcp auth out of harness to server
