--- packages/harness/src/core/mcp/auth/index.ts
@@ -1,23 +1,0 @@
-export {
-  buildMcpAuthorizationUrl,
-  completeMcpAuthorization,
-  createMcpOAuthClient,
-  ensureMcpClientRegistered,
-  isMcpAuthRequired,
-  resolveMcpAuth,
-} from './mcpDcr';
-export type {
-  CompleteMcpAuthorizationResult,
-  McpAuthRequiredResult,
-  McpAuthResolvedResult,
-  ResolveMcpAuthResult,
-} from './mcpDcr';
-export {
-  DEFAULT_MCP_ACCESS_TOKEN_TTL_SECONDS,
-  MCP_OAUTH_CALLBACK_PATH,
-  mcpAuthorizationServerMetadata,
-  mcpAuthorizationServerOrigin,
-  mcpClientInformation,
-  mcpOAuthCallbackUrl,
-} from './mcpOAuthHelpers';
-export { validateRedirectUris } from './redirectUri';
