+++ packages/server/package.json
@@ -29,6 +29,7 @@
     "@hono/node-server": "^2.0.11",
     "@hono/swagger-ui": "^0.2.1",
     "@hono/zod-openapi": "^1.5.1",
+    "@modelcontextprotocol/sdk": "^1.29.0",
     "@truefoundry/utils-core": "workspace:*",
     "better-sqlite3": "^13.0.2",
     "hono": "^4.12.34",
