+++ packages/server/src/apis/mcpOAuth.ts
@@ -1,12 +1,8 @@
 import { OpenAPIHono, type RouteHandler } from '@hono/zod-openapi';
-import {
-  completeMcpAuthorization,
-  extractErrorLogFields,
-  McpConnectionError,
-  type IOAuthClientStore,
-  type IOAuthTokenStore,
-} from '@truefoundry/utils-core/core';
+import { extractErrorLogFields, McpConnectionError } from '@truefoundry/utils-core/core';
 import type { Logger } from 'winston';
+import { completeMcpAuthorization } from '../mcp/auth/mcpDcr';
+import type { IOAuthClientStore, IOAuthTokenStore } from '../mcp/auth/types';
 import { mcpOAuthCallbackRoute } from '../routes/mcpOAuthRoutes';
 
 export interface McpOAuthRouterDeps {
