+++ packages/server/src/mcp/auth/mcpDcr.ts
@@ -12,47 +12,29 @@ import type {
   OAuthClientMetadata,
   OAuthTokens,
 } from '@modelcontextprotocol/sdk/shared/auth.js';
+import { McpConnectionError, McpDcrConfigurationError } from '@truefoundry/utils-core/core';
 import { randomBytes } from 'node:crypto';
-import type { IOAuthClientStore, OAuthClientRecord } from '../../auth/IOAuthClientStore';
-import type { IOAuthTokenStore } from '../../auth/IOAuthTokenStore';
-import { isOAuthAccessTokenUsable } from '../../auth/oauthToken';
-import { McpConnectionError, McpDcrConfigurationError } from '../../errors';
 import {
+  isOAuthAccessTokenUsable,
   mcpAuthorizationServerMetadata,
   mcpAuthorizationServerOrigin,
   mcpClientInformation,
   mcpOAuthCallbackUrl,
   oauthTokensToOAuthToken,
 } from './mcpOAuthHelpers';
-
-export interface McpAuthResolvedResult {
-  headers: Record<string, string>;
-}
-
-export interface McpAuthRequiredResult {
-  authUrl: URL;
-}
-
-export type ResolveMcpAuthResult = McpAuthResolvedResult | McpAuthRequiredResult;
+import type {
+  CompleteMcpAuthorizationResult,
+  IOAuthClientStore,
+  IOAuthTokenStore,
+  McpAuthRequiredResult,
+  OAuthClientRecord,
+  ResolveMcpAuthResult,
+} from './types';
 
 export function isMcpAuthRequired(result: ResolveMcpAuthResult): result is McpAuthRequiredResult {
   return 'authUrl' in result;
 }
 
-/** Result of a successful `completeMcpAuthorization` (token saved, pending cleared). */
-export interface CompleteMcpAuthorizationResult {
-  serverId: string;
-  /** FE post-OAuth landing URL from the original authorize call. */
-  redirectUrl: string | null;
-}
-
-/**
- * Prefer `client_secret_post` (confidential). Authorization servers that only issue public
- * clients or only accept other methods can reject that; retry once without the field (same as
- * servicefoundry outbound DCR), then surface. Not a loop.
- * When the token response omits `expires_in` (allowed by RFC 6749), use a 1h TTL so
- * the saved token is not treated as already expired on the next `resolveMcpAuth`.
- */
 async function registerMcpClientWithAuthMethodFallback(params: {
   authorizationServerUrl: string;
   metadata: AuthorizationServerMetadata;
@@ -63,7 +45,6 @@ async function registerMcpClientWithAuthMethodFallback(params: {
 }): Promise<OAuthClientRecord> {
   const { authorizationServerUrl, metadata, clientMetadata, mcpServerName, authorizationEndpoint, tokenEndpoint } =
     params;
-
   let fullInfo;
   try {
     fullInfo = await registerClient(authorizationServerUrl, {
@@ -79,11 +60,7 @@ async function registerMcpClientWithAuthMethodFallback(params: {
       );
     }
     try {
-      // Omit method so the authorization server applies its default.
-      fullInfo = await registerClient(authorizationServerUrl, {
-        metadata,
-        clientMetadata,
-      });
+      fullInfo = await registerClient(authorizationServerUrl, { metadata, clientMetadata });
     } catch (secondError: unknown) {
       throw new McpConnectionError(
         `Failed to dynamically register OAuth client for MCP server '${mcpServerName}'`,
@@ -92,14 +69,12 @@ async function registerMcpClientWithAuthMethodFallback(params: {
       );
     }
   }
-
   if (!fullInfo.client_id) {
     throw new McpConnectionError(
       `Authorization server for MCP server '${mcpServerName}' returned a client response without client_id`,
       400,
     );
   }
-
   return {
     server: {
       authorizationEndpoint,
@@ -120,30 +95,23 @@ export async function createMcpOAuthClient(params: {
   clientName: string;
 }): Promise<OAuthClientRecord> {
   const { mcpServerUrl, mcpServerName, redirectUri, clientName } = params;
-
   const { authorizationServerUrl, authorizationServerMetadata: metadata } = await discoverOAuthServerInfo(mcpServerUrl);
-
   if (!metadata?.registration_endpoint) {
     throw new McpDcrConfigurationError(
       `MCP server '${mcpServerName}' has no DCR support (missing registration_endpoint); auth.type: dcr is misconfigured for this server`,
     );
   }
-
   if (!metadata.authorization_endpoint || !metadata.token_endpoint) {
     throw new McpDcrConfigurationError(
       `Authorization server for MCP server '${mcpServerName}' is missing authorization_endpoint or token_endpoint`,
     );
   }
-
-  // Only reject when the AS explicitly advertises PKCE methods that omit S256.
-  // Missing/omitted advertisement → still try S256 (many servers support it but do not publish).
   const pkceMethods = metadata.code_challenge_methods_supported;
   if (pkceMethods && !pkceMethods.includes('S256')) {
     throw new McpDcrConfigurationError(
       `Authorization server for MCP server '${mcpServerName}' advertises PKCE methods without S256`,
     );
   }
-
   return registerMcpClientWithAuthMethodFallback({
     authorizationServerUrl,
     metadata,
@@ -170,14 +138,12 @@ export async function ensureMcpClientRegistered(params: {
   if (existing) {
     return existing;
   }
-
   const client = await createMcpOAuthClient({
     mcpServerUrl: params.mcpServerUrl,
     mcpServerName: params.mcpServerName,
     redirectUri: mcpOAuthCallbackUrl(),
     clientName: params.clientName,
   });
-
   await params.mcpServerStore.saveClient({ id: params.serverId, record: client });
   return client;
 }
@@ -189,7 +155,6 @@ export async function buildMcpAuthorizationUrl(params: {
   mcpServerUrl: string;
   mcpServerName: string;
   clientName: string;
-  /** FE post-OAuth landing URL (not the OAuth redirect_uri). */
   redirectUrl?: string;
 }): Promise<URL> {
   const client = await ensureMcpClientRegistered({
@@ -199,20 +164,15 @@ export async function buildMcpAuthorizationUrl(params: {
     mcpServerName: params.mcpServerName,
     clientName: params.clientName,
   });
-
   const state = randomBytes(32).toString('base64url');
   const redirectUri = mcpOAuthCallbackUrl();
-  const resource = resourceUrlFromServerUrl(params.mcpServerUrl);
-
-  // MCP SDK always sends PKCE S256; it only rejects when metadata lists methods that omit S256.
-  // When we stored null (AS did not advertise), reconstructed metadata omits the field → try S256.
   let started: Awaited<ReturnType<typeof startAuthorization>>;
   try {
     started = await startAuthorization(mcpAuthorizationServerOrigin(client.server), {
       metadata: mcpAuthorizationServerMetadata(client.server),
       clientInformation: mcpClientInformation(client.client),
       redirectUrl: redirectUri,
-      resource,
+      resource: resourceUrlFromServerUrl(params.mcpServerUrl),
       state,
     });
   } catch (error: unknown) {
@@ -220,7 +180,6 @@ export async function buildMcpAuthorizationUrl(params: {
       cause: error instanceof Error ? error : undefined,
     });
   }
-
   await params.tokenStore.savePendingAuthorization({
     state,
     id: params.serverId,
@@ -228,7 +187,6 @@ export async function buildMcpAuthorizationUrl(params: {
     codeVerifier: started.codeVerifier,
     redirectUrl: params.redirectUrl ?? null,
   });
-
   return started.authorizationUrl;
 }
 
@@ -239,16 +197,13 @@ export async function resolveMcpAuth(params: {
   mcpServerUrl: string;
   mcpServerName: string;
   clientName: string;
-  /** FE post-OAuth landing URL (not the OAuth redirect_uri). */
   redirectUrl?: string;
 }): Promise<ResolveMcpAuthResult> {
   const nowMs = Date.now();
   const token = await params.tokenStore.getToken({ id: params.serverId });
-
   if (token && isOAuthAccessTokenUsable(token.expiresAt, nowMs)) {
     return { headers: { Authorization: `Bearer ${token.accessToken}` } };
   }
-
   if (token?.refreshToken) {
     const client = await params.mcpServerStore.getClient({ id: params.serverId });
     if (client) {
@@ -263,17 +218,13 @@ export async function resolveMcpAuth(params: {
         await params.tokenStore.saveToken({ id: params.serverId, token: saved });
         return { headers: { Authorization: `Bearer ${saved.accessToken}` } };
       } catch {
-        // Refresh failed → drop the unusable token and fall through to full re-authorize.
+        // Refresh failure falls through to a fresh authorization.
       }
     }
   }
-
-  // Only reached when there is no usable/refreshable token. Clear the stale row before re-auth
-  // (successful refresh returns above and never deletes).
   if (token) {
     await params.tokenStore.deleteToken({ id: params.serverId });
   }
-
   const authUrl = await buildMcpAuthorizationUrl({
     tokenStore: params.tokenStore,
     mcpServerStore: params.mcpServerStore,
@@ -286,12 +237,6 @@ export async function resolveMcpAuth(params: {
   return { authUrl };
 }
 
-/**
- * OAuth callback: exchange `code` for tokens. Route must already reject IdP `error` params.
- * Claims pending state atomically (`consumePendingAuthorization`) so duplicate callbacks lose
- * the race; `mcpServerUrl` for the RFC 8707 `resource` comes from that pending row.
- * On `invalid_client`, clears stored client + token so the next authorize re-registers.
- */
 export async function completeMcpAuthorization(params: {
   tokenStore: IOAuthTokenStore;
   mcpServerStore: IOAuthClientStore;
@@ -299,13 +244,10 @@ export async function completeMcpAuthorization(params: {
   code: string;
 }): Promise<CompleteMcpAuthorizationResult> {
   const nowMs = Date.now();
-
-  // Consume first so a concurrent/duplicate callback cannot redeem the same state.
   const pending = await params.tokenStore.consumePendingAuthorization({ state: params.state });
   if (!pending) {
     throw new McpConnectionError('Unknown or expired OAuth state', 400);
   }
-
   const client = await params.mcpServerStore.getClient({ id: pending.id });
   if (!client) {
     throw new McpConnectionError(
@@ -313,19 +255,16 @@ export async function completeMcpAuthorization(params: {
       400,
     );
   }
-
-  const { codeVerifier } = pending;
-  if (codeVerifier === null) {
+  if (pending.codeVerifier === null) {
     throw new McpConnectionError('Pending authorization is missing PKCE code_verifier; re-run authorize', 400);
   }
-
   let tokens: OAuthTokens;
   try {
     tokens = await exchangeAuthorization(mcpAuthorizationServerOrigin(client.server), {
       metadata: mcpAuthorizationServerMetadata(client.server),
       clientInformation: mcpClientInformation(client.client),
       authorizationCode: params.code,
-      codeVerifier,
+      codeVerifier: pending.codeVerifier,
       redirectUri: mcpOAuthCallbackUrl(),
       resource: resourceUrlFromServerUrl(pending.mcpServerUrl),
     });
@@ -341,11 +280,9 @@ export async function completeMcpAuthorization(params: {
       cause: error instanceof Error ? error : undefined,
     });
   }
-
   await params.tokenStore.saveToken({
     id: pending.id,
     token: oauthTokensToOAuthToken(tokens, nowMs, null),
   });
-
   return { serverId: pending.id, redirectUrl: pending.redirectUrl };
 }
