+++ packages/server/tests/db/mcpServerStoreContractSuite.ts
@@ -2,8 +2,8 @@
  * Backend-agnostic behavioural contract for IMcpServerStore.
  * Runs under jest against a fresh store per test (see backend test files).
  */
-import type { OAuthClientRecord } from '@truefoundry/utils-core/core';
 import type { IMcpServerStore } from '../../src/db/mcpServerStore';
+import type { OAuthClientRecord } from '../../src/mcp/auth/types';
 import type { McpServerManifest } from '../../src/schemas/mcpServer';
 
 const TENANT = 'default';
